// webapp/src/types/api/GetEntitiesResponse.ts
// This file is auto-generated by `cargo run --bin export_types`
// Do not edit manually.

// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Entity } from "../database/Entity";

export type GetEntitiesResponse = { entities: Array<Entity>, total_count: number, page: number, limit: number, total_pages: number, };